+++ src/database/schema.ts
@@ -218,9 +218,14 @@ const schemaModel = {
 			})
 		)[0]?.id;
 
-		if (existingService && !schemaId && await versionExists(trx, existingService.id, service.version)) {
-			const message = `Schema [${existingService.name}] and version [${service.version}] already exist in registry. `
-				+ `You should not register different type_defs with same version.`;
+		if (
+			existingService &&
+			!schemaId &&
+			(await versionExists(trx, existingService.id, service.version))
+		) {
+			const message =
+				`Schema [${existingService.name}] and version [${service.version}] already exist in registry. ` +
+				`You should not register different type_defs with same version.`;
 			throw new PublicError(message, null);
 		}